use size_t for last video/audio frames and avoid static casts

+++ libbase/FLVParser.h
@@ -17,7 +17,7 @@
 // Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 //
 
-// $Id: FLVParser.h,v 1.4 2007/05/05 15:52:50 strk Exp $
+// $Id: FLVParser.h,v 1.5 2007/05/05 16:19:18 strk Exp $
 
 // Information about the FLV format can be found at http://osflash.org/flv
 
@@ -246,10 +246,10 @@ private:
 	FLVAudioInfo* _audioInfo;
 
 	/// Last audio frame returned
-	uint32_t _lastAudioFrame;
+	size_t _lastAudioFrame;
 
 	/// Last video frame returned
-	uint32_t _lastVideoFrame;
+	size_t _lastVideoFrame;
 
 	/// Audio stream is present
 	bool _audio;
